Notes & fixed test case

+++ firethorn-core/src/test/java/uk/ac/roe/wfau/firethorn/adql/query/atlas/CastFunctionTestCase.java
@@ -210,7 +210,7 @@ extends AtlasQueryTestBase
 
             new ExpectedField[] {
                 new ExpectedField("binsize",  AdqlColumn.AdqlType.LONG,    0),
-                new ExpectedField("binvalue", AdqlColumn.AdqlType.INTEGER, 0),
+                new ExpectedField("binvalue", AdqlColumn.AdqlType.LONG, 0),
                 }
             );
         }
